pyproject.toml
[tool.poetry]
name = "bus-ticket-bot"
version = "0.1.0"
description = "Bus Ticketing Bot (Telegram-Based)"
authors = ["Hossein Naseri <dev.h.naseri@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
peewee = "^3.18.3"
pytelegrambotapi = "^4.29.1"
khayyam = "^3.0.17"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
